Moved all polybench header files into directory polybench. Adapted polyhedral benchmarks to refer to those headerfiles.
