+++ micro-benchmarks/DRB055-jacobi2d-parallel-no.c
@@ -11,10 +11,10 @@
 #include <string.h>
 #include <math.h>
 /* Include polybench common header. */
-#include <polybench.h>
+#include "polybench/polybench.h"
 /* Include benchmark-specific header. */
 /* Default data type is double, default size is 20x1000. */
-#include "jacobi-2d-imper.h"
+#include "polybench/jacobi-2d-imper.h"
 /* Array initialization. */
 
 static void init_array(int n,double A[500 + 0][500 + 0],double B[500 + 0][500 + 0])
